+++ docs/resources/changelog.md
@@ -6,7 +6,7 @@ Components with the <sl-badge type="warning" pill>Experimental</sl-badge> badge 
 
 _During the beta period, these restrictions may be relaxed in the event of a mission-critical bug._ 🐛
 
-## Next
+## 2.0.0-beta.60
 
 - Added React examples and CodePen links to all components
 - Changed the `attr` in experimental `<sl-mutation-observer>` to require `"*"` instead of `""` to target all attributes